Fix web compile errors: nullable promotion + num/double types
- settings: assert non-null on closure-captured 2FA error before AuthErrorInline
- budget/ui/reports: .clamp() returns num, coerce to double for
  ProgressBar/LinearProgressIndicator/FractionallySizedBox and chart maxY

+++ lib/widgets/ui.dart
@@ -158,7 +158,7 @@ class ProgressBar extends StatelessWidget {
         color: p.bgInput,
         child: FractionallySizedBox(
           alignment: Alignment.centerLeft,
-          widthFactor: fraction.clamp(0.0, 1.0),
+          widthFactor: fraction.clamp(0.0, 1.0).toDouble(),
           child: Container(
             decoration: BoxDecoration(
               color: gradient == null ? (color ?? AppColors.primary) : null,
